feat(#448): remove redundant package
