+++ src/main/java/org/eolang/jeo/AssembleMojo.java
@@ -31,7 +31,6 @@ import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
-import org.eolang.jeo.improvement.TranslationBytecodeFootprint;
 
 /**
  * Converts EO to bytecode.
