+++ src/main/java/org/eolang/jeo/TranslationBytecodeFootprint.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.jeo.improvement;
+package org.eolang.jeo;
 
 import com.jcabi.log.Logger;
 import java.io.IOException;
@@ -30,9 +30,6 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Collection;
 import java.util.Collections;
-import org.eolang.jeo.Details;
-import org.eolang.jeo.Representation;
-import org.eolang.jeo.Translation;
 import org.eolang.jeo.representation.JavaName;
 
 /**
